compaction: move owned ranges filtering to base class
Move the token filtering logic down from cleanup_compaction
to regular_compaction and class compaction so it can be
reused by other compaction types.

Create a _owned_ranges_checker in class compaction
when _owned_ranges is engaged, and use it in
regular_compaction::make_sstable_reader to filter
partitions based on the owned ranges.

Note that _owned_ranges_checker was defined as mutable
so that make_partition_filter that is declared as const
can call `_owned_ranges_checker->belongs_to_current_node`
that may modify _owned_ranges_checker.  This will be simplified
in the next patch.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>