compaction_manager: keep owned_ranges_ptr in table_state
When perform_cleanup adds sstables to cleanup_sstable_set,
also save the owned_ranges_ptr in the table_state so
it could be used by other compaction types like
regular, reshape, or major compaction.

When the exhausted sstables are released, check
if cleanup_compaction_set is empty, and if it is,
clear also the owned_ranges_ptr.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>