storage_service: run_replace_ops: use topology nodes
Use topology to get to track
nodes to sync with all the way to
node_ops_cmd_heartbeat_updater and send_node_ops_cmd.

Also, base the list of nodes to sync with
on the token_metadata get_endpoint_to_host_id_map_for_reading
thta lists all normal token owners, rather than on the nodes
in gossip, as the token_metadata is the source of truth
for node operation as in all other ops.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>