host_id_or_endpoint: resolve: return node_ptr
Use std::variant<host_id, inet_address> to keep
either of them and resolve using topology
to a node_ptr.

Reflect that also from token_metadata::parse_host_id_and_endpoint
that gets a string describing either a host_id or inet_address
and returns a node_ptr if found.

Exceptions are thrown if the string is invalid
or if the node is not found in topology.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>