repair: node_ops_cmd_request: support node_ops_cmd_v2
With node_ops_cmd_v2, use the node index as the
inet_address in the existing node_ops_cmd_request fields
and send an additional dictionary mapping each node index
to a pair of <host_id, inet_address>.

Provide a set of getters that extract either the legacy
field from the request if the dictionary isn't provided,
or use the dictionary to translate the raw addresses,
that contain node indexes into the real node address.

A later patch, will use the host_id in the dictionary
to look up nodes in storage_service::node_ops_cmd_handler
and both the host_id and inet_address to populate
new nodes in topology for the replace and bootstrap case.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>