abstract_replication_strategy: add in_natural_endpoints
Currently, effective_replication_map::do_get_ranges accepts
a functor that traverses the natural endpoints of each token
to decided whether a token range should be returned or not.

This is done by copying the natural endpoints vector for
each token.  However, other than special strategies like
everywhere and local, the functor can be called on the
precalculated inet_address_vector_replica_set in the
replication_map and there's no need to copy it for each call.

in_natural_endpoints passes a reference to the function
down to the abstract replication strategy to let work either
on the precalculated inet_address_vector_replica_set or
on a ad-hoc vector prepared by the replication strategy.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>