topology: optimize compare_endpoints
This function is called on the fast data path
from storage_proxy when sorting multiple endpoints
by proximity.

This change uses precalculated node location metrics
to use numeric comparison instead of boolean logic.
The change is not entirely backward compatible
since the previous implementation returned 0
in cases we didn't care to order the nodes
while the new code may return std::string_ordering::less
or greater in this case.  But that shouldn't matter
and the test_topology_compare_endpoints unit test
verifies the ordering result only where it matters.

This change reduces the compare_endpoints footprint
to a point it is no inlined in topology::sort_by_proximity.

Previously, the functions binary size in bytes were as follows:
  based on objdump -d output
  with clang version 15.0.7 (Fedora 15.0.7-1.fc37)

sort_by_proximity=896
compare_endpoints=8752
total 9648

with this optimization:
sort_by_proximity=1520

so the improvement is by 84%

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>