topology: add node location metric
To be used for optimizing compare_endpoints
in the next patch.

To precalculate the location metric for each endpoint
we keep track of each dc and dc/rack radix in
corresponding maps and then we enumerate all
endpoints in the order they are inserted.

The node location metric is computed by
(dc_radix << 48) | (rack_radix << 32) | node_radix
so that nodes can be sorted by relative proximity to each other
based on a proposal by Avi Kivity <avi@scylladb>.

The order within each dc/rack doesn't matter
as long as we can determine that nodes are the same.

Similarly for ordering different racks within the same dc,
the relative order between them is not important, as long
as nodes in the same rack sort closely to each other.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>