storage_service: add node_ops_ctl class to formalize all node_ops flow
All node operations we currently support go through
the same basic flow and may add some op-specific logic
around it.

1. Select the nodes to sync with (this is op specific).
2. send prepare req
3. hearbeat updater
4. perform the body of the node operation
5. send done
--
on any error: send abort

node_ops_ctl formalizes all those steps and makes
sure errors are handled in all steps, and
the error causing abort is not masked by errors
in the abort processing, and is propagated upstream.

Some of the printouts repeat the node operation description
to remain backward compatible so not to break dtests
that wait for them.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>