storage_service: standardize error messages when aborting node ops
Alwasy start with an error describing the failure
that contains the current_exception (that is propagated
after sending the abort requests).

After send all abort requests log a message at info level
that abort operations completed. At the moment it's printed
in various log levels as warn/error, but it's not an error by itself
maybe info is even too high for it and it could be demoted
to debug.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>