sstables: generation_type: add new_generation function
new_generation gets an optional previous generation.
If absent, new_generation initializes numeric generations
to 1, and if engaged, new_generation returns a new
generation with the following generation number.

In the future, with uuid-based generations, the
function could be used to generate a new random
uuid based on the optional parameter.

It will be up to the caller, e.g. replica::table or
sstables manager to decide which kind of generation to
create.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>