storage_service: get_ignore_dead_nodes_for_replace: make static and rename to parse_node_list
Let the caller pass the string to parse to the function
rather than the function itself get to it via _db.local().get_config()
so it could be used as a general purpose function.

Make it static now that it doesn't require an instance.

Rename to `parse_node_list` as that's what the function does.
It doesn't care if the nodes are to be ignored or something else
(e.g. removed), they only need to be in token_metadata.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>