gossiper: map handlers by msg_addr
We should identify nodes by their host_id
rather than endpoint, that may change or may transfer
to a different node when replacing a node reusing the old
node address.

This change starts modifying the internal gossiper
state to use msg_addr that may hold the node's host_id
in addition to its address after the node is upgraded to
a recent version.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>