gossiper: get_endpoint_id from rpc::client_info
Get the host_id and address auxiliary attributes
from client_info.  If host_id is null (when communicating
with an outdated node), get the host_id from topology.
We can assume that the node will be in topology
as all known host_id:s are loaded by storage_service::join_cluster
from system.peers.

If the host_id is not found, get_host_id_and_endpoint
prints a warning and throws a runtime_error so the
caller can be sure that the host_id is available
when it returns successfully.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>