gossiper: add_saved_endpoint: add host_id parameter
The caller (storage_service) loads the host_id
at this point so it can simply pass it on to
add_saved_endpoint, rather than the latter
look up the host_id indirectly in token_metadata.

Note that storage_service can pass the loaded
endpoint_state too. Can be considered as a followup.

The host_id, addr are used also to update the address map.

Make sure that the host_id is valid (present and non-null)
since we rely on that later-on to be able to identify
nodes by their host_id.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>