row_cache: abort on exteral_updater::execute errors
Currently the cache updaters aren't exception safe
yet they are intended to be.

Instead of allowing exception to escape execute(),
make it noexcept in the pure virtual external_updater_impl.
Implementation that are able to provide strong exception
guarantees in execute() should still throw nothing,
but make the exception available via get_exception()
so that it can be propagated from cache::update/invalidate.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>