compaction_backlog_tracker: reset backlog manager in move-assignment
Currently the move-assugnment operator exchanges the moved-to
backlog manager with the moved-from manager, and deregisters
the moved-to object from its backlog manager.

This is not required sinve the move-assignment operator
is called only from compaction_manager::register_backlog_tracker
which re-registers the moved-to tracker with the same manager.

The moved-from tracker is deregitered from its manager
anyhow when is destroyed, so this had no effect on it anyhow.

To make this path clearer, just exchange the move-to
manager with nullptr to leave it in a valid, unregistered,
state.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>