sstable_set: keep impl as shared_ptr
For atomic cache update, we need to isolate
updates to the sstable list.
Hole the cache update_sem provide write/write
isolation, but the mechanism we use for read/write
isolation which to clone the whole sstable_set
is too coarse-grained, requiring us to refresh
the table compound sstable set when any of the
compaction groups' sstable sets is updated.

Keeping a shared impl in the sstable set
allows us to update just this set without
requiring rebuild of the whle table's sstable set.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>