compaction_backlog_tracker: operator=: do not change tracker registration
Handling of the tracker's registration should be up
to the caller.

Currently, the move assignment operator exchanges
the assigned tracker's backlog manager pointer with the
moved tracker's manager; it unregisters the
assigned tracker, but doesn't register it to the moved
tracker's manager, so the `_manager` member is
just left broken. Luckily, this is benign, since
the move asasignment is done only from
`compaction_manager::register_backlog_tracker` which
proceeds with registering the assigned tracker
to cs.backlog_tracker.

Fix this simply by keeping the assigned tracker
registration intact.  This will become useful in a
followup patch that will update the tracker in place
with a modified clone of it, during cache updates.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>