database: update_keyspace_on_all_shards: abort on error
Currently errors from any shard from db.update_keyspace
or from the respective notification will leave the
database shards in an inconsistent state.

It is better to abort in this case than to ignore the failure
since scylla can then reload the schema from scratch.

We don't do that for create/drop keyspace since
this affects only that keyspace that is either not used yet
(on create) or not used anymore (in the drop case).
FWIW, the error is now propagated upstream.

Refs #3389

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>