gossiper: fold update_live_endpoints_version into replicate_live_endpoints_on_change
We want to propagate any change to _live_endpoints
to all shards.  Currently we just update the `_live_endpoints_version`
and `replicate_live_endpoints_on_change` propagtes the
change some undetermined time in the future.

To rely on `_live_endpoints` for gossiper::is_alive,
that may be called on any shard, we want to propagate
the change to all shards as soon as it happens.

The `_endpoint_update_semaphore` is acquired,
as required by `replicate_live_endpoints_on_change`,
using `lock_endpoint_update_semaphore`.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>