gossiper: add get_live_endpoint_states
So the caller will not have to call is_alive for
each endpoint.

Use in migration_manager::schedule_schema_pull

Note that unlike `get_endpoint_states`, `get_live_endpoint_states`
returns a copy of the endpoint_state map.  This is okay since
it is not on the hot path.

A following series, supporting the management of endpoint_state
map by host_id is palanned to keep a lw_shared_ptr<endpoint_state>
so it can be index both by host_id and inet_address.
With that in mind, copy the map will just copy the lw_shared_ptr,
without copying the actual endpoint_state.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>