raft_address_map: return future from replicate and add_or_update_entry
and wait on them.

`_replication_fiber` made into an optional shared_future
so we can both wait on it on stop()
and return a future to the caller.

Propagate the future up the stack in `replicate_add_or_update_entry`
and `add_or_update_entry` and wait for it to resolve in the
different call sites.

The motivation for this change is to use the address_map
directly from gossiper and ensure all shards are updated
when entries are added or updated.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>