table: add table_sstable_set
Replace the table compound_sstable_set with
a table_sstable_set that takes advantage of
the compaction groups mutual disjointness
to implement a stateless sstable set that doesn't
need to be refreshed any time any of the
compaction groups' sstable sets is updated.

The table_sstable_set creates a compound_sstable_set
for each compaction groups it is working on
but otherwise, it gurantees no atomicity of selection
across compaction groups.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>