table: add table_sstable_set
Replace the table compound_sstable_set with
a table_sstable_set that takes advantage of
the storage_group:s mutual disjointness
to implement a stateless sstable set that doesn't
need to be refreshed every time any of the
compaction groups' sstable set is updated.

The table_sstable_set selector creates a
compound_sstable_set for each storage group it is working on.
Otherwise, it guarantees no atomicity of selection
across storage groups.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>