system_keyspace: templatize execute_cql_query
Rename `execute_cql` to `execute_cql_query` and make
it a template, accepting a `const std::ranges::range&`
for the data values, corresponding to a similar
change in the cql3::query_processor.

A following patch will add a use case passing
a ref to `std::vector<data_value>` of values.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>