storage_service: do not update peer info for this node
system_keyspace had a hack to skip update_peer_info
for the local node, and then to remove an entry for
the local node in system.peers if `update_tokens(endpoint, ...)`
was called for this node.

This change unhacks system_keyspace by considering
update of system.peers with the local address as
an internal error and fixing the call sites that do that.

Fixes #16425

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>