gossiper: add_saved_endpoint: keep heart_beat_state if ep_state is found
Currently, when loading peers' endpoint state from system.peers,
add_saved_endpoint is called.
The first instance of the endpoint state is created with the default
heart_beat_state, with both generation and version set to zero.
However, if add_saved_endpoint finds an existing instance of the
endpoint state, it reuses it, but it updates its heart_beat_state
with the local heart_beat_state() rather than keeping the existing
heart_beat_state, as it should.

This is a problem since it may confuse updates over gossip
later on via do_apply_state_locally that compares the remote
generation vs. the local generation, so they must stem from
the same root that is the endpoint itself.

Also, call on_internal_error_noexccept if add_saved_endpoint is called
while there's already an endpoint_state for the node with
non-zero generation, as it indicates it was received from
gossip and should not be updated with local information
loaded from system.peers. This will abort in testing or just
print an error in production and do nothing more.

Fixes #16429

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>