repair: do_rebuild_replace_with_repair: use source_dc only when safe
It is unsafe to restrict the sync nodes for repair to
the source data center if it has too low replication factor
in network_topology_replication_strategy, or if other
nodes in that DC are ignored.

Also, this change restricts the usage of source_dc
to `network_topology` and `everywhere_topology`
strategies, as with simple replication strategy
there is no guarantee that there would be any
more replicas in that data center.

Fixes #16826

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>