repair: do_rebuild_replace_with_repair: use source_dc only when safe
It is unsafe to restrict the sync nodes for repair to
the source data center if we cannot guarantee a quorum
in the data center with network-topology replication strategy.

Also, this change restricts the usage of source_dc
to `network_topology` and `everywhere_topology`
strategies, as with simple replication strategy
there is no guarantee that there would be any
remaining replicas in that data center.

Fixes #16826

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>