repair: do_rebuild_replace_with_repair: use source_dc only when safe
It is unsafe to restrict the sync nodes for repair to
the source data center if we cannot guarantee a quorum
in the data center with network-topology replication strategy.

Also, this change restricts the usage of source_dc
to `network_topology` and `everywhere_topology`
strategies, as with simple replication strategy
there is no guarantee that there would be any
remaining replicas in that data center.

When source_dc is given, we distinguish between several cases.

1. It is considered unsafe to use source_dc if number of nodes
   lost in that DC (replaced/rebuilt node + additional ignored nodes)
   is greater than 1, or it has 1 lost node and rf=1 in the DC.

2. If the source_dc arg is forced, as with the new
   `nodetool rebuild --force <source_dc>` option,
   we use it anyway, even if it's considered to be unsafe.
   A warning is printed in this case.

3. Otherwise, we look for an alternative source datacenter,
   that has not lost any node. If such datacenter is found
   we use it as source_dc for the keyspace, and log a warning.

4. If no alternative dc is found, then:

   4.a If the source_dc option is user-provided (as with `nodetool
   rebuild`), An exception is thrown and nodetool will fail.

   4.b Otherwise, like with `nodetool repair`, we log a warning
   and fall back to using replicas from all nodes in the cluster

Fixes #16826

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>