repair: do_rebuild_replace_with_repair: reject source_dc if rf=0 or no live nodes in the dc
The source_dc is given by the user when invoking the rebuild api.
Rather than falling back to using the whole cluster
when we cannot rebuild from source_db, throw an error
instructing the user to use a different source_dc or no source_dc
altogether.

When we cannot ensure quorum for rebuild, just log a warning,
but use this source_dc anyway.

Note that the source_dc is not provided by the user for
replace, just for rebuild, so we fallback to using the whole cluster
for replace in case we cannot ensure quorum in the source_dc
since the user has no control of it.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>