compaction_descriptor: enable_garbage_collection: make cloning of input sstable_set explicit
To make sure that the snapshot taken would be immutable.
A following patch replaces the table's _sstables set
with a new `table_sstable_set` class that is not immutable
like today's compound_sstable_set, so we need to take
a copy of the current sstable_set as snapshot.

Currently the sstable_set is copied since `enable_garbage_collection`
gets it by value and then moves it, but this is quite subtle.
This patch makes the copy explicit by accepting the input
sstable_set by reference and then emplacing a copy of it
into `all_sstables_snapshot`.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>