table: add shard_local_storage_group_at
Adds a getter for storage_group that checks if the
requested storage_group, given its index, is owned
by the local shard and returns pointer to it if its
owned, or nullptr otherwise.

Rename `storage_group_for_token` respectively
to `shard_local_storage_group_for_token` since
it too now returns nullptr if the token is not
owned by the current shard.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>