table_sstable_set: clone efficiently
Currently, to clone a table_sstable_set, we create
a partitioned_sstable_set of all sstables, to take
advantage of the distjointness of sstables in different
storage_groups.  But, building the new interval set
is rather expensive.

Instead, in this patch, we clone the storage_groups,
one at a time, and skipping storage_groups that aren't
owned by the current shard.

Each storage_group clones its compaction_groups
and each of them, its sstable_sets respectively.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>