tablet_storage_group_manager: precalculate my_host_id and _tablet_map
The node host_id never changes, so get it once,
when the object is constructed.

A pointer to the tablet_map is taken when constructed
using the effective_replication_map_ptr and it is
updated whenever the e_r_m changes, using a newly added
`update_effective_replication_map` method.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>