locator: topology: node: set endpoint_dc_rack as part of indexing
Adding the node* to the right dc/rack
maps is part of indexing it, so apply
that mapping as the last stage of index_node
and only set the node's _dc_rack member
after it's indexed.

Following patch will make that exception-safe
by using native datacenter and rack objects
that would be set in topology atomically,
when ready.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>