locator: topology: add class location
Define a location class and keep it
as a member in class node.

Also, keep a _default_location member in
class topology pointing to default datacenter
and rack objects that are allocated when topology
is constructed.

Those are kept around even if empty since
_default_location points to them.

When cloning the topology, create respective
default datacenter and rack on the destination
topology, and copy the datacenter and rack tree
to the cloned objects.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>