cql_server: connection: process_on_shard: drop permit parameter
It is currently unused in `process_on_shard`, which
generates an empty service_permit.

The next patch may call process_on_shard in a loop,
so it can't simply move the permit to the callee
and better hold on to it until processing completes.

In the future we may want to solve it differently
by passing the permit to `process_on_shard` and
on to the `process_fn`, but the latter will have
to return it back if it returns a bound_msg
(and then the permit will need to be wrapped in
a foreign_ptr).

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>