replication strategies: add per-strategy traits
Refactor traits like the strategy type,
is it local, is it per table, or does it use tablets
into abstract_replication_strategy_traits
and derive a per-strategy traits class off of it
to prepare the per-strategy traits.

This is done since in several places we create
a full blown replication strategy just to check
its traits.

Next patch will add a topology& parameter to
create_replication_strategy, which is not easily
available in the mentioned call sites, so those
should only create the strategy traits that just
process the replication parameters without
doing heavier calculations (network_topology_strategy
in particular).

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>